web/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.4.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
drf-spectacular==0.28.0
drf-writable-nested==0.7.2
environs==11.0.0
geopy==2.4.1
gunicorn==23.0.0
Pillow==11.1.0
pgvector==0.4.0
psycopg[binary,pool]==3.2.1
setuptools==78.1.0
